Use more fine-grained API for everything
With this we no longer send any PUT requests to /api/things.

This uses a positively awful lock file mechanism to avoid data races that would
corrupt or even crash the server. It's very fragile, and definitely not a good
way to handle things, so I really need to come up with something smarter than
this!