Store data as JSON in MongoDB
This just stores all the data for a given user as JSON encoded as a string. This
is obviously not the approach we want to take long-term.

We also implement a pretty wonky locking mechanism to avoid the case where
concurrent updates causes outdated information to be written back into the
database.