Switch from Draft.js to Slate.js for the editor
I'm finding Draft really unintuitive to use, and the documentation is quite
sparse. Slate seems to have a much nicer API, though the documentation also
isn't great, but it should be a lot easier to learn how to do more complex
things with it.

I briefly experimented with rolling my own contenteditable-based editor, but
even just making newlines behave correctly was a huge pain, so I gave up on that
pretty quickly.