Get rid of automatic garbage collection
This does not take references or parents into account, so we risk getting rid of
items that the user actually want. For this reason, we'll remove it for
now. Eventually we will probably also want to add a list of all items in the
system, and then this won't be useful anyway.