Refactor tree module so it doesn't directly depend on representation
A separate module now defines the underlying representation; all functions in
the main tree module depend on this module. This should make it easier when we
want to add new functionality that requires changing the underlying
representation. We also did a bunch of cleanup in the Tree module while
implementing this change.