Implement a functional demo
We've copy-pasted a lot of code here. Guess someone else will have to clean that
up later :)

Also, the way we've implemented the data for the demo is a bit inelegant, but
it'll do for now. If it turns out that we want to go back and edit it quite
frequently, then we may want to store the data in a more sensible way.