Simplify special behavior for parents and references
We now reuse the parents and references trees that would normally appear as
children of the given item in the zoomed-in view. This means that we now in
principle only need a single TreeContext in the program, which will be very nice
in the future, I think. We also get limited drag-and-drop support for free.