Add basic example of CORS over localhost
We can now host dist/static/ on localhost:8085 and dist/server/ on
localhost:8080. Then, we can log in through localhost:8080, and then
use the app on localhost:8085.

This implementation is meant as a prototype; we need to make the static
and server domains configurable, add instructions for deployment,
make login work from the static site, and clean up the code required to
do all this.

This is the first step towards having the static resources be hosted on
a separate domain.