Implement addChild in terms of insertChild
We would like to implement everything in terms of insertChild and removeChild,
so that we can store both children and parents directly in the
datastructure. This will make functions like parent() and remove() constant with
respect to the total number of items in the database. (Of course, remove() will
still be linear in the number of parents, but this doesn't really matter in
practice.)