Fix bug with removing thing
Sometimes, nodes get detached from the tree without being removed. When this
happened to a node being removed by Tree.removeThing, it would throw an error.

Note that TypeScript would have caught this error if we had enabled strict mode,
which we should definitely do.

We should also consider garbage collection for nodes in the tree, so detached
nodes are removed.