Implement tagged connections on the client-side only
We want to do much more with this feature, including actually syncing to the
server! This is a good start, though.

I had to separate refreshChildren back out from genericRefreshChildren, and it
seems to have caused some problems with focus when creating a sibling (though
I'm not sure why). Eventually, we probably want to have only a very barebones
genericRefreshChildren, and use manual refreshing almost everywhere for
performance.