Use string IDs for connection types
This is done for the same reason that we use this ID format for things. If we
just used an incrementing numeric ID, two different clients could easily end up
creating a connection with the same ID, which would cause everything to break.

(Currently, connections are not actually stored on the server directly, so this
is not an issue right now.)