Fix broken synchronization between clients
There was a pretty glaring bug in updateLocalState that I'm amazed I hadn't
caught until now. It would update an outdated tree with the new state, causing
the new tree to be incorrect.

The fix for this involves nesting a bunch of state updating functions, which I
don't feel super good about, but it seems to work.