Fix serious bug in how children are calculated by 'GET /state'
We forgot to filter to just connections by the correct user, so the children
would include *all* connections by all users.

I couldn't really figure out how to write this code in a clean way, so we ended
up with... this. Maybe it's worth considering switching to a SQL database.