Send information about connections to server
This moves the endpoint 'PUT /api/things/:thing' to 'PUT /state/things/:thing',
and makes it so that the client must mention connections by name.

One notable advantage of this approach is that we can now store tags on the
server, although we currently don't read them on the client.