Add API endpoint to update multiple items at once
This fixes a bug where the server would sometimes receive updates out of
order. This would cause it to try to insert a connection to an item that would
only be created in the next request, causing the database to return an error,
and thus creating an incorrect state where a connection was missing.

Now, related updates are sent together, and we can choose their order.