Simplify how focus is handled
Previously, when a context editor was blurred, that editor would be responsible
for unfocusing the item in the tree.

Now, focus is handled globally in response to events to the outermost div
element.

This is mainly important for how the toolbar is handled. When a user clicks a
button on the toolbar, we *don't* want the "logical
focus" (i.e. T.focused(tree)) to correspond to the DOM focus.

We were previously using a workaround, but this workaround did not work properly
with touch events. This new method is cleaner, and works fine on mobile too!