Add untested macOS archive
We switched to using ".tar.gz" instead of DMG. I'm not familiar enough with
macOS to know whether this thing can even be run at all. I think DMG is the
typical way to distribute applications on macOS, but we can't make that from
Linux.

Unfortunately, I don't have a macOS machine, so I don't know if it's possible to
run the contents of the ".tar.gz" archive -- or how to do so.