Add "Storage" wrapper around some server operations
The idea is that we want to be able to run Thinktool without a server (i.e. on
the desktop), using a different method to store data. So we have abstracted the
parts of the server that deal with storing data into a new interface, which we
can later have multiple implementations of.