Use 'json_content' as source of truth in Database module
This is really inefficient, because we convert from JSON to strings, send that
over the wire, and then convert from strings back to JSON on the
client. However, this commit means that we can now remove the old 'content'
column, and we no longer have to worry about synchronizing updates between
columns.