When fetching full state from server, do more work in DB
I'm not sure if this is really a good idea. It makes queries slower on the
database, but lets us really clean up our application code.

Eventually, I think we might just want to store children directly as JSON
together with the items themselves. Then we should be able to really optimize
this query.