Fix bug where you couldn't click on toolbar items on macOS
I'm still not sure what exactly causes this, but by experimenting a bit, I
managed to figure out that just calling ev.preventDefault() in onMouseDown seems
to have fixed it.

It's odd that this bug occurs in all browsers on macOS, but not on Linux or
Windows. I don't know why.