Use more robust types for basic data structure
Previously, we would use some types like '{[id: string]: Value}', where looking
up certain keys would actually yield 'undefined'. TypeScript accepts this, even
though it's not correct.

We now explicitly annotate such types with '| undefined'. That being said, in
some places, we have no way of handling such errors, so we just throw instead.

This did fix a bug where editing the content of a non-existent item would crash
the application. This now creates the item instead. This bug was relevant when
the user manually typed a link, and then openeded it and tried to edit it.

Not updating the version of the desktop application was done intentionally. See
9628a4e for more information about this.