Improve logging for user login
We used the old logging to debug a weird problem with logins, but that turned
out to be related to the version of Node being used. The problem was fixed in
the same (somewhat inaccurately named commit) that added the logging, namely
4246c83.

That being said, it's probably a good idea to print something when a user can't
log in or when the client sends a weird request. There's no reason to log their
password here, though!

We should probably add similar logging to the rest of the server, to get a
better idea of whether clients are sending us invalid requests.