Fix desktop build
Previously we received this error from GitHub Actions:

    > thinktool-desktop-client@0.1.2 build-static /home/runner/work/diaform/diaform/src/desktop
    > cp ../../dist/static/* build && cp index.html build && rm -f build/icon.png

    cp: -r not specified; omitting directory '../../dist/static/blog'

But the root cause is that we use a really awkward build system where we build
static assets, and then snatch those at the last minute for the desktop client.

In this commit, we've done some work to partially improve the situation. We now
use custom HTML files for the desktop client, and we specifically pick out the
static files that we want.

We still rely on hijacking the old build pipeline to build the client.

We also still use a dumb hack where we build the client to dist/static, and then
remove all other files so our CI/CD system uploads the right files. I can't
remember why I originally decided to do this, but we'll have to fix it in a
different commit.