Don't make caller of 'actionsWith' manually pass in node
This should always be the focused node, so there should be no reason why
we can't just find this node directly in 'actionsWith'. I believe we had
this hack because the focus system was previously quite flaky, and would
randomly break for no reason. To the best of my knowledge, focus is now
reasonably robust.

We still have the problem where some actions allow the node to be null
and others don't.