Add explicit 'execute' to Actions module
This seems like a better interface than what we had previously. Now, a
consumer of this module can check whether an action can be executed; if
it can, then they can execute that action.

Actions can now be referred to directly, instead of being represented as
functions, which will be useful for attaching metadata to them.