Split up GitHub Actions workflows
Often we just want to redeploy static resources.

Our current setup requires us to also recompile the desktop client and
update the API server over SSH. This is too heavy.

With this approach, we split up the workflow into three different
workflows. These new workflows trigger on different branches, namely
'prod-api', 'prod-static' and 'prod-desktop'.

This will require us to manually push each branch when we want to
trigger updates. I don't know if this tradeoff is worth it, but let's
try it out.