Merge "Create" with input in thing selection popup and prefer match
Instead of having a separate result for creating a new item, this is now
merged into the search box itself.

We also select matches by default, instead of selecting the new item. We
automatically select the new item if it's the only available item. If
the query is changed such that more matches become available, then the
selection is reset.