Support multiple newsletters in 'send.py' script
This script now takes an argument describing which newsletter to send.
It takes advantage of the new database schema (see 01aaddb1) to store
information about each newsletter separately.

Since I'm in the process of switching from plain SMTP to using the
SendGrid API, I haven't implemented the sending itself in this commit --
obviously, we'll need to do that before this is really useful!