Copy sent email to file instead of sending copy to cc@thinktool.io
This is done mostly to reduce our SendGrid usage. It also reduces the
amount of spam sent to cc@thinktool.io. We mostly had this functionality
as a way to check that the generated output was correct. It's sufficient
to just check whatever was sent last, and verify that it looks right.