Add background and animation for bullet point on hover
This looks nicer and should hopefully make it more obvious that you can
click on bullet points.

In order to be able to animate them with CSS, we need to put the SVGs
inline.

We just updated the bullet points in the outline for now – links will
have to come later. It looks like we forgot to remove the
bullet-link-*.svg files when we stopped using them, but now we're using
those again while waiting on links to also be animated.