Use published NPM packages when building static resources
The main change in this version is that we now no longer build the
'client' and 'shared' packages when we update the static web resources.
The new script to build static resources is called
'tools/build-web.sh', and just uses 'npm ci' for dependencies – it does
not use local dependencies.

This means that we will have to manually publish '@thinktool/client'
(and therefore '@thinktool/shared') in the future, before deploying
static resources.