Add Actions workflow for building desktop app as artifact
This will be useful for testing. Getting the right environment set up to
build the desktop clients is really annoying, so it'd be nice if we
could just build these on GitHub Actions, and then test the client
manually. But we may not want to actually deploy them into production,
so the 'deploy-desktop' workflow isn't really good enough.