Remove references to pages from UI
I've decided to remove pages because I haven't really been able to do
anything interesting with them.

This commit just removes pages from the toolbar and the stylesheet, so
they are invisible to users, but the application still stores
information about pages, which we probably want to remove as well.