Turn GitHub Actions workflows into manual dispatches, rather than per-branch
Previously we had a bunch of dummy branches which were only used to
trigger these workflows. GitHub recently added the 'workflow_dispatch'
trigger, which lets us trigger these workflows manually instead.

The reason for doing this over, say, having a single 'production' branch
is that rebuilding static assets is *much* faster than rebuilding the
desktop client, so we don't want to waste a bunch of time constantly
rebuilding the desktop client when we just want to push a minor update.