Improve appearance of "Demo" button, and replace one of them
We replaced the large "Go to demo" button in the main context with a
signup button. If people aren't going to click the big red button in
the toolbar, they're probably not going to click it the second time
they see it either. This should make it more explicit that you *can*
actually sign up for an account.