Handle most other keyboard shortcuts
Inserting a link still doesn't work; everything else should.

For this commit, we removed the special-case for handling arrow keys,
since we no longer need to use the old hack. Instead, we pass
information about whether we are at the first/last line of the item,
and then let the 'onKeyDown' handler take care of any special behavior.