Partially fix some of the wonkiness
This is an ugly commit, but basically there are two important parts:

1. We were using a stale context inside of the keyboard event handler
because it did not respond to changes in the 'onKeyDown' prop. We get
around this by doing a trick where we save the prop in a ref and then
update it whenever it changes. Our event handler doesn't change ever.

2. We weren't responding to changes in the state. Now we are.

I think this editor still has a few problems, but it's *mostly* at feature
parity with the old one, to the best of my knowledge.