Improve consistency of internal links in editor
They now look closer to internal links in unfocused items. The behavior
is subtly different; see the TODO.

This approach of splitting up InternalLink is arguably a hack, but we
would like to make the dependencies of InternalLink more explicit
anyway. So if we refactor this in the future, we should make
InternalLink look more like InternalLink_ -- not the other way around.