Specify version of Node and server dependencies
Previously, we were using Node 14 in the development image but 13 in the
production image, which just seems like a bad idea.

We now use Node version 12, which is the LTS version, in both images.

We also pinned the packages used on the server, which were previously
set to 'latest' in most cases.