Extract searching functionality to its own module
Even though we generally prefer a more functional style in this project,
we use a class for the search module. This is because we can't easily
represent the state with an immutable object, so our only alternative
would be to provide a collection of procedures that mutate some object.
But in that case, I prefer making only the procedures available, and
keeping the shared state hidden. Classes provide a convenient syntax for
this.