Don't regenerate index on each character typed
And clean up code in general.

We now create the index when the popup is opened, and then don't update
it unless the state is changed (which shouldn't happen in practice).

However, it turns out the the search itself, not the indexing, is what's
slow. This is mostly a problem for longer queries.

Even slow searches only take a few seconds, so this is something that we
could probably fix with a nicer UI. As mentioned in the comment, this
would require making the search asynchronous.