Hardcode URLs for splash images in client package
Logically, I feel like these images should be part of the package
itself, but that's non-trivial to do. This is at least an improvement
over having an unstated requirement on having these files on whatever
server you're running the app from.

Updated version to a '-dev', so I hopefully remember to bump it before
merging into master.