Add 'demo.html' for use with Parcel
The idea is that we can now run 'parcel serve demo.html', and Parcel
will automatically build everything for us and start a development
server.

I think this is probably how we were supposed to be using Parcel all
along...

I want to integrate this style into the development and build workflow,
but this commit is just a little proof of concept for how it should
work.

We need to add the other pages too, and then we need to update the build
scripts to use this method instead of what we're doing currently.