Fix bug where invalid queries would hang server
This is a pretty serious bug. If there was an error in the query that
we sent to the database, then the client used for that query would never
be freed. Eventually, the entire client pool would be exhausted, and
then the server would simply stop responding to requests.

We fix this by adding a 'withClient' wrapper that has some rudimentary
error handling, and using that everywhere. (We actually had something
like this previously, but were only using it in some places.)

Obviously, we would prefer to not be sending invalid queries to the
database in the first place.