Use Parcel when building client for production
This is much cleaner than what we were doing before. The web package is
now responsible for building itself, and then we just grab the output
and put it in dist/static/ when we're ready to deploy.

For some reason we had to change a bunch of percentages in the SCSS
files to decimals, otherwise they would be set to 1%. I found this
relevant StackOverflow post, but I have no idea why this is happening:
https://stackoverflow.com/q/58853844