Handle invalid links when getting textual content
Previously when an item contained a link to an item that didn't exist in
the search popup, a warning would be printed in the console, and that
link would be replaced by an empty item.

Now, invalid links are shown explicitly anywhere 'contentText' is used,
such as in the search popup.