Merge GitHub Actions workflows for web and desktop
Previously, we sometimes wanted to update the web client without also
updating the desktop client. We've now made the desktop build
sufficiently automated for this to not be relevant. Whenever we update
the web client, we now also want to update the desktop client.

Technically there is a problem with this workflow where the downloads
page is sometimes updated before the new files are actually uploaded.