Fix crash when removing a cyclic connection from a child to itself
This was caused by the fact that we didn't consider this case in 'removeChild',
so we would keep the stale data of the "child" around even after updating the
"parent" -- but these two items can be the same.

The result was that the item would mention a child connection that didn't exist,
and looking up this connection would cause a crash.