Fix bug where parents of focused item would not be updated
Specifically, we forgot to update these items when adding it as a child of
another item or when removing it as a child of another item.

Our current data model is really inelegant, since it stores the so-called
'otherParentsChildren' for all nodes, when this is actually something that we
only care about for the focused item. But I haven't deciced where I want to go
with this in the future yet, so maybe we'll end up using it.

Anyway, this fixes an annoying bug!