Fix broken build due to mismatched Parcel vs. custom build config
This reverts commit 83c439b9d2e494ab382bad1c424d5ccb42ad5665.

Basically: When we build the web client, we use Parcel. Parcel wants URLs inside
the Sass stylesheets to be relative references to an actual file on disk. But
when we build the other static resources from our custom
script (build-static.sh), we just build the stylesheets as-is, so here we need
absolute URLs relative to the website root.

When we added the icon in the commit above, we added an aboslute URL, which
breaks the Parcel build. We'll have to rethink how we do this in the future, but
this commit fixes the immediate problem.