Export React components from 'client' module
Previously we exported functions that would then mount those React components on
an element with a predefined ID. This seems like a cleaner approach, but it
means that we'll have to update the places where this module is used.

We want to avoid using two versions of React (one from this module and another
from the user), so we added React to 'peerDependencies' instead. I'm not
entirely confident that this approach will work, but let's see!