Build demo with Next.js instead of Parcel
This *mostly* works pretty well. There are only two things that are slightly
inelegant about this:

1) We need to use resources from outside this package (namely 'app.css' in
'@thinktool/client' and the shared resouce 'icon.png'). We just manually build
these files into the 'public/' directory. I can't think of a better approach
here.

2) Parcel won't let us remove 'demo.html' since we have a link to it, so we just
added a placeholder instead. This is a hack, but it's only there until we're
done with Parcel, and the build still works.