Add landing page to Next.js project
We had to build the stylesheets manually here, because Next.js doesn't let us
import our own stylesheets in the same way that we can import stylesheets from
packages for some reason. The long-term solution might be to use CSS modules,
but I don't want to do that in this commit.