Start working on "goals" system
The idea is that we want to give the users some tasks, which are automatically
marked as finished. This might be a nice way to introduce the user to the most
important features in a more interactive manner than the current tutorial.

This commit technically works, but it's very WIP.