Manually copy CSS from '@thinktool/client' instead of using 'import'
This reverts commit c7e17121138b9eb745ed62daf8f69c1e2e7c5894.

The old method of importing CSS was annoying when doing development. When we
have 'npm link'ed the client package from the web package, then Next.js starts
complaining about the fact that we're importing a local stylesheet globally on a
page, which is not supported.

This method works whether or not packages are linked.