Hook up goals from actions individually
I think we abstracted a little too quickly with our work in the last
commit. Let's just have the goals module listen to a specific set of events and
then trigger those as needed.

Note that te way we're handling state here is a little ugly and not
React-y. It's not obvious to me that there isn't potential for race conditions
or other state desync issues with all this 'context.setWhatever' business.