Set up skeleton for message dispatch architecture
The idea is that we want to centralize all state updates in a single
place. Since we need to trigger various actions from a large number of places in
the UI, a message passing style seems like a good fit.

We want to slowly convert the UI to send messages instead of triggering state
changes directly. Then we want to subscribe to those messages in a central
place, and handle them there.