Start refactoring actions and add unit test for 'new'
The idea is that actions should be pure functions that simply step the state
forward, instead of being allowed to manipulate state directly.

Currently, we're just doing this refactoring inside the actions module itself by
extending its interface. However, we'll probably want to completely remove the
old stateful interface and then change how we use actions elsewhere in the
program.