Convert updates to run asynchronously
The idea is that we want to call out to some external behavior, for example
opening a popup to prompt for an item, in the middle of an action. One way to do
this is using 'async'.

Of course, the problem with this is that the state may change while the async
code is running. We need to manually ensure that this doesn't happen.