Also implement "insert-link" as update but break "undo" and "forum"
This required us to use a pretty hacky approach. See the comment. We return a
special result when a link should be returned, and then apply that. The fact
that this actually works seems to mostly be a coincidence.

We also broke these other actions so we could get rid of the old implementation
mechanism immediately. I'll do those in the following next commit.