Start refactoring click callbacks and add tests for opened links
This is only the first bit of a refactoring. We want to make it possible to
write unit tests for clicking on the bullets next to items. In order to do this,
we would like to export a very straight-forward 'click' function (and
'altClick') from the item module.

This is only implemented for links in this commit, which results in a bunch of
extra complexity.