Take callback for registering callback to manage selection locally
This is still pretty bad, but it is an improvement. Rather than calling a
callback on the global 'Context' object that registers a callback to be called
later, we now call a callback that is passed as a prop, and then the consumer of
the 'Editor' module must register the callback that's passed to that callback to
wherever it needs to go.