Keep our own editor as state inside editor component
Previously, we took the editor state from our parent, and then kept our own
cached state for ProseMirror. Now that we have all the information that we care
about in the 'Editor' interface, we would rather keep our component-local state
in such an object.

We can then just generate the state for ProseMirror, without needing to keep any
state specifically for that.

It's weird that we take a *different* Editor through our props, but that we then
only care about the content on it. We should probably merge these concepts
somehow.