Configure test directories explicitly
This avoids an issue where tests that were complied to 'dist/' would still be run
by Jest. We can put tests both in 'tests/', and directly in 'src/'. So far, we've
been separating our tests, but I think we'll want to keep them together with the
code under test in the future.