Fix compilation error for 'app.html' during development
The old version worked fine when we built the project for production. But when
using 'npm run dev', we would get an error when trying to access
'/app.html'. This is apparently related to the stylesheet import. We already
made a similar change for 'demo.html' in a previous commit.