Get rid of nasty setTimeout hack in popup component
Now, the actions are no longer async. Instead, each action updates the app state
such that it has an open popup. Together with the popup, we store a "pure
callback" (i.e. type App => App, more or less) that is called by the popup
component in response to user input, and which is then used to update the state.

The nice thing about this approach is that the popup system no longer depends on
user interaction. For example, we could write a function like this pseudocode:

    let app = initialAppState();
    app = selectFirstItem();
    app = moveCursorToEndOfSelectedItem();

    app = insertLink(); // <- This changes the app state so a popup is open

    app = setPopupQuery("This is a new item.");
    app = movePopupSelectionUp();
    app = selectSelectedItemInPopup(); // <- This uses the "pure callback" to
                                       //    update the app state

In theory, we should no longer need the `input` function to be returned from the
popup component; we can do all that as a normal function called directly from
the actions module. I just haven't gotten around to removing it.