Move the searching itself to a worker
Getting Webpack, TypeScript, Next.js and Web Workers to cooperate was very
annoying, but it looks like I made it work. One thing that I struggled with a
bunch was getting "ReferenceError: self is not defined".

This was because I originally tried to construct the worker in the
top-level. This would cause it to be executed by Next.js on the server side,
which we don't want. As a simple workaround, we now instead create the worker
the first time it's actually used, so it happens only on the client side.