Memoize outline and search popup items
This addresses lag while typing in search popup.

We don't need to rerender the whole outline when the popup is open. Arguably,
this use of `React.memo` is very slightly hacky, since the outline probably
shouldn't know about the popup to begin with; we should have a better way of
determining whether or not to rerender.

In the popup module, we just memoize the items in the list, which helps a lot
with avoiding lag when typing.