Use effects system for handling search
This also causes the popup to be populated with results immediately, instead of
waiting for user input. (If text has been selected before opening the popup.)

Previously, we handled search inside the <Popup> component, which was obviously
not a good approach.

The problem is that searching is asynchronous: We send a message to an external
module when we want to search for something, and then receive results back
later. We can't do this while keeping our functions functional and our control
flow simple.

The solution is to use a similar approach to how we handle other effects like
"open URL" in the actions module. Functions that cause a search to be
initialized now return an effect, and we intercept that effect elsewhere. This
is set up in the main module.