Reflect focus changes in editor view
Oops, I was a bit quick with that last commit. When an editor lost focus, we
weren't actually blurring the editor itself. I don't think ProseMirror provides
a way to do this, so we have to manually call the 'blur' method.

This looks more hacky than it really is; ProseMirror and TypeScript just can't
quite agree on what's happening, so we have to do some type coercion.