Use Yarn Workspaces to manage dependencies
We haven't updated the documentation or build system for this change yet.

The idea is that this should let us avoid a lot of the annoying plumbing around
keeping package versions in sync during development (namely, keeping the version
of React used by `@thinktool/client` and `@thinktool/web` in sync when using the
latter package to test the former).