Use 'style-loader' to eliminate CSS file from client distribution
Now, the consumer of the '@thinktool/client' package no longer needs to worry
about including the correct CSS.

The downside of this is that Next.JS doesn't like the code that we're now
generating (since it tries to run it on the server-side, where style-loader
can't inject the style tags). However, we can work around that my using dynamic
import statements to ensure that the module is only loaded on the client
side. We were doing a more hacky version of this previously, so this isn't
actually any worse than before.