Show appropriate icon for connecting actions
Previously we would decide the icon based only on the editor state. So for
example, after pressing Alt-P in an editor to select a parent, the link icon
would be shown in the focused search bar.

We now use more complex logic for deciding which icon to show. In particular, we
change the icon to something appropriate when an action is run that activates
the search bar.