Fix issue with server build by upgrading all dependencies
I think we upgraded the dependency for @johv/miscjs in the server
package but not elsewhere, so this dependency would get stuck inside
the server's node_modules, which we don't copy over when building the
server. Fixed by simply upgrading all dependencies.

Arguably this is an issue with our over-complicated build process, not
a dependency issue. This will probably come up again in the future.