Expand <Item> component to include the item's subtree
This is a pretty huge improvement to the code quality, since we can now
represent an entire tree of items as plain data (using the newly introduced
`ItemData`). This entire tree can then be rendered directly using <Item>, which
is a pure component.

We have introduced this new component and use it inside <ExpandableItem> in the
main module. But we want to go through the main module and remove a bunch of old
code that should now no longer be necessary.