Remove unnecessary non-pure components from main module
Now that we have this new <Item> module, we can clean up a bunch of code in the
main module.

Now, the entire state of the outline is represented as a singe ItemData, which
is a fully transparent, pure JSON object. This object is constructed once for
the root item (and, for now, once for each of its parents) from the stateful
representation of the graph.

And item events are now thrown all the way back up to a central callback, which
is defined at the same time as the root item. This callback never changes, so we
don't need to worry about changes in our event handlers in the leaves.

This is a huge improvement to code quality! We're actually starting to use React
the way it was intended to be used.