Show indicator when server can't be contacted after changes received
This is a pretty rare scenario: We successfully get a message from the server
that some items have changed, but then we can't contact the server to actually
load those changes.

It will be more useful to implement the same error message in cases where we
can't send data to the server when making changes to items.