Pull in server-side state when user selects "Keep Remote"
We fixed a bug in 'transformFullStateResponseIntoState', where we previously
would mutate the 'empty' state value(!), which would cause issues when using
this value multiple times. Oops!

For now, we just re-initialize it manually each time. This function is kinda
ugly in general, mostly because it's one of the few parts of this application
where we actually care about performance somewhat.