Preserve editor state when moving items around
This is a rather ugly implementation, but having a working test of this
functionality should give a lot more confidence about future improvements to the
tree module.

The nice thing about relying only on the outline module in our tests is that we
have a lot of freedom to change major aspects of the implementation in the
future without breaking these tests. (In constast to the tests in, say,
`tree.test.ts`, which are much less robust and will most likely need to be
deleted eventually.)