Fix interaction between editor focus and popups
Not sure when this was introduced. Due to this bug, a popup would not appear
when triggering an action that should open a popup from a new item that had not
received any edits.

I'm not sure exactly what caused this specific bug, but it seems to be related
in some way to our unnecessary double-handling of focus changes in
`<ProseMirror>` as well as the fact that items did not relinquish their focus
while the popup is open.

This also fixes another bug where the cursor would occationally jump around
after changing focus from one editor to another.