Remove `.html` suffix from paths on website
This is partly because it looks nicer, and partly because getting Next.js to
behave in a reasonable way with the `.html` suffixes there is pretty much
impossible. In particular, Next.js has different behavior when running in
development versus production. So let's just let Next.js decide what our paths
should look like.

The changes in this commit will be deployed together with a change in the
configuration for the server such that redirects are handled correctly.