Add development target for client package
Previously, we had to link the `client` and `web` packages manually, and then
run `web` to see our changes to `client`. Now we can run

    $ yarn webpack --config webpack.dev.js

and then run some static server that serves the `dev.html` file to run a version
of the demo directly in the `client` package.

from the `src/client/` directory, and then see the changes live. This is a much
nicer workflow.